changed to .jsx files
