--- src/components/PhotoGallery/PhotoGallery.js
@@ -1,12 +1,0 @@
-.import React, { Component } from 'react'
-
-export default class PhotoGallery extends Component {
-    render() {
-        return (
-            <React.Fragment>
-                <PG_PhotoExpanded></PG_PhotoExpanded>
-                <PG_Thumbnails></PG_Thumbnails>
-            </React.Fragment>
-        )
-    }
-}
