+++ src/components/PhotoGallery/PhotoGalleryPhotoExpanded.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from "react";
+
+export class PhotoGalleryPhotoExpanded extends Component {
+	render() {
+		return <div></div>;
+	}
+}
+
+export default PhotoGalleryPhotoExpanded;
