+++ src/components/PhotoGallery/PhotoGalleryThumbnail.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from "react";
+
+export class PhotoGalleryThumbnail extends Component {
+	render() {
+		return <div></div>;
+	}
+}
+
+export default PhotoGalleryThumbnail;
